working sshkit with config script

+++ lib/tungsten.rb
@@ -19,8 +19,46 @@ module Tungsten
     end
 
     def load!
-      eval(File.open(options[:config_file], 'r').read)
-      run_dsl!
+      if File.exists?(options[:config_file])
+        eval(File.open(options[:config_file], 'r').read)
+      end
+    end
+
+    def servers_iteration(&block)
+      roles.each do |role|
+        @servers.each do |server|
+          if server.roles.include?(role)
+            SSHKit::Backend::Netssh.configure do |ssh|
+              ssh.ssh_options = server.ssh_options
+            end
+            yield server, role
+          end
+        end
+      end
+    end
+
+    def execute!
+      puts "Execute server commands"
+      servers_iteration do |server, role|
+        execution_commands = @execution_commands[role]
+        if execution_commands
+          execution_commands.yield(server)
+        else
+          puts "No execution commands found for role: #{role}"
+        end
+      end
+    end
+
+    def install!
+      puts "Execute server installation"
+      servers_iteration do |server, role|
+        installation_commands = @installation_commands[role]
+        if installation_commands
+          installation_commands.yield(server)
+        else
+          puts "No installation commands found for role: #{role}"
+        end
+      end
     end
   end
 end
